refactor: 멀티 모듈 구조 도입 및 공통 모듈 분리
- 프로젝트를 멀티 모듈 구조(common, auth-server)로 개편
- 공통 예외 처리, 응답 클래스, 상수 및 유틸리티를 common 모듈로 이관
- auth-server가 common 모듈을 의존하도록 build.gradle 수정
- 프로젝트 루트로 Gradle 래퍼 및 설정 파일 이동
- 패키지 이동에 따른 import 경로 수정 및 컴포넌트 스캔 범위 확장
