--- auth-server/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'auth-service'
