--- auth-server/src/main/java/com/truve/platform/auth/service/common/exception/CustomException.java
@@ -1,15 +1,0 @@
-package com.truve.platform.auth.service.common.exception;
-
-import lombok.Getter;
-
-@Getter
-public class CustomException extends RuntimeException {
-
-	private final ErrorCode errorCode;
-
-	public CustomException(ErrorCode errorCode) {
-		super(errorCode.getMessage());
-		this.errorCode = errorCode;
-	}
-
-}
