--- auth-server/src/main/java/com/truve/platform/auth/service/common/support/BaseEntity.java
@@ -1,15 +1,0 @@
-package com.truve.platform.auth.service.common.support;
-
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-
-@Getter
-@MappedSuperclass
-public abstract class BaseEntity extends BaseTimeEntity{
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	protected Long id;
-}
