--- auth-server/src/main/java/com/truve/platform/user/service/common/constants/AuthProvider.java
@@ -1,15 +1,0 @@
-package com.truve.platform.user.service.common.constants;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum AuthProvider {
-
-	LOCAL,
-	GOOGLE,
-	NAVER,
-	;
-
-}
