--- auth-server/src/main/java/com/truve/platform/user/service/common/exception/ApiAdvice.java
@@ -1,44 +1,0 @@
-package com.truve.platform.user.service.common.exception;
-
-import java.util.Arrays;
-
-import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import com.truve.platform.user.service.common.response.ErrorResponse;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@RestControllerAdvice
-@RequiredArgsConstructor
-public class ApiAdvice {
-	private final ApplicationEventPublisher applicationEventPublisher;
-	@ExceptionHandler(Exception.class)
-	public ResponseEntity<ErrorResponse.ErrorData> internalServerError(Exception e) {
-		applicationEventPublisher.publishEvent(
-			e.getMessage()
-		);
-
-		return ErrorResponse.error(HttpStatus.INTERNAL_SERVER_ERROR, "서버에러입니다. 백엔드팀에 문의하세요.");
-	}
-
-	@ExceptionHandler(CustomException.class)
-	public ResponseEntity<ErrorResponse.ErrorData> customException(CustomException e) {
-		return ErrorResponse.error(e.getErrorCode().getStatus(), e.getErrorCode().getMessage());
-	}
-
-	@ExceptionHandler(MethodArgumentNotValidException.class)
-	public ResponseEntity<ErrorResponse.ErrorData> methodArgumentNotValidException(MethodArgumentNotValidException e) {
-		var details = Arrays.toString(e.getDetailMessageArguments());
-		var message = details.split(",", 2)[1].replace("]", "").trim();
-
-		return ErrorResponse.error(HttpStatus.BAD_REQUEST, message);
-	}
-
-}
