+++ auth-server/src/main/java/com/truve/platform/user/service/controller/EmailController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.truve.platform.user.service.common.response.ApiResult;
+import com.truve.platform.common.response.ApiResult;
 import com.truve.platform.user.service.domain.dto.request.EmailRequest;
 import com.truve.platform.user.service.service.EmailService;
