+++ auth-server/src/main/java/com/truve/platform/user/service/domain/entity/User.java
@@ -1,8 +1,8 @@
 package com.truve.platform.user.service.domain.entity;
 
-import com.truve.platform.user.service.common.constants.AuthProvider;
-import com.truve.platform.user.service.common.constants.UserRole;
-import com.truve.platform.user.service.common.support.BaseEntity;
+import com.truve.platform.common.constants.AuthProvider;
+import com.truve.platform.common.constants.UserRole;
+import com.truve.platform.common.support.BaseEntity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
