+++ common/src/main/java/com/truve/platform/common/constants/AuthProvider.java
@@ -1,0 +1,15 @@
+package com.truve.platform.common.constants;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum AuthProvider {
+
+	LOCAL,
+	GOOGLE,
+	NAVER,
+	;
+
+}
