+++ common/src/main/java/com/truve/platform/common/constants/UserRole.java
@@ -1,0 +1,8 @@
+package com.truve.platform.common.constants;
+
+public enum UserRole {
+	ADMIN,
+	MEMBER,
+	ARTIST,
+	;
+}
