+++ common/src/main/java/com/truve/platform/common/exception/CustomException.java
@@ -1,0 +1,15 @@
+package com.truve.platform.common.exception;
+
+import lombok.Getter;
+
+@Getter
+public class CustomException extends RuntimeException {
+
+	private final ErrorCode errorCode;
+
+	public CustomException(ErrorCode errorCode) {
+		super(errorCode.getMessage());
+		this.errorCode = errorCode;
+	}
+
+}
