+++ common/src/main/java/com/truve/platform/common/response/ErrorResponse.java
@@ -1,0 +1,30 @@
+package com.truve.platform.common.response;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ErrorResponse {
+
+	private HttpStatus status;
+	private String message;
+
+	public static ResponseEntity<ErrorData> error(HttpStatus status, String message) {
+		return ResponseEntity.status(status).body(ErrorData.of(status.series().name(), message));
+	}
+
+	@Getter
+	@AllArgsConstructor
+	public static class ErrorData {
+		private String code;
+		private String message;
+
+		public static ErrorData of(String code, String message) {
+			return new ErrorData(code, message);
+		}
+	}
+}
