+++ common/src/main/java/com/truve/platform/common/support/BaseEntity.java
@@ -1,0 +1,15 @@
+package com.truve.platform.common.support;
+
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+
+@Getter
+@MappedSuperclass
+public abstract class BaseEntity extends BaseTimeEntity{
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	protected Long id;
+}
