+++ common/src/main/java/com/truve/platform/common/support/BaseTimeEntity.java
@@ -1,0 +1,24 @@
+package com.truve.platform.common.support;
+
+import java.time.LocalDateTime;
+
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+
+@Getter
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public abstract class BaseTimeEntity {
+
+	@CreatedDate
+	protected LocalDateTime createdAt;
+
+	@LastModifiedDate
+	protected LocalDateTime updatedAt;
+
+}
