+++ settings.gradle
@@ -1,0 +1,3 @@
+rootProject.name = 'truve-backend'
+include ':common'
+include ':auth-server'
